move top rows heap and heap sort into sorters package
